requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-crispy-forms==2.4
django-jalali==7.4.0
django-jalali-date==2.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
Khayyam==3.0.17
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
